feat(ticket-687): implement technical seo features
