+++ src/app/sitemap.ts
@@ -1,0 +1,37 @@
+import { MetadataRoute } from 'next';
+import { BRAND } from '../lib/core/brand';
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  return [
+    {
+      url: `${BRAND.url}/`,
+      lastModified: new Date(),
+      changeFrequency: 'weekly',
+      priority: 1,
+    },
+    {
+      url: `${BRAND.url}/docs`,
+      lastModified: new Date(),
+      changeFrequency: 'weekly',
+      priority: 0.8,
+    },
+    {
+      url: `${BRAND.url}/philosophy`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly',
+      priority: 0.8,
+    },
+    {
+      url: `${BRAND.url}/privacy`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly',
+      priority: 0.5,
+    },
+    {
+      url: `${BRAND.url}/terms`,
+      lastModified: new Date(),
+      changeFrequency: 'monthly',
+      priority: 0.5,
+    },
+  ];
+}
